fix report paths
