+++ tools/ci_connector_ops/ci_connector_ops/pipelines/commands/airbyte_ci.py
@@ -4,13 +4,11 @@
 
 """This module is the CLI entrypoint to the airbyte-ci commands."""
 
-from pathlib import Path
 from typing import List
 
 import click
 from ci_connector_ops.pipelines import github, main_logger
 from ci_connector_ops.pipelines.bases import CIContext
-from ci_connector_ops.pipelines.consts import LOCAL_REPORTS_PATH_ROOT
 from ci_connector_ops.pipelines.utils import (
     get_current_epoch_time,
     get_current_git_branch,
@@ -18,7 +16,6 @@ from ci_connector_ops.pipelines.utils import (
     get_modified_files_in_branch,
     get_modified_files_in_commit,
     get_modified_files_in_pull_request,
-    upload_to_gcs,
 )
 from github import PullRequest
 
@@ -50,45 +47,6 @@ def get_modified_files(
         else:
             return get_modified_files_in_branch(git_branch, git_revision, diffed_branch, is_local)
     return get_modified_files_in_branch(git_branch, git_revision, diffed_branch, is_local)
-
-
-def render_report_output_prefix(ctx: click.Context) -> str:
-    """Render the report output prefix for any command in the Connector CLI.
-
-    The goal is to standardize the output of all logs and reports generated by the CLI
-    related to a specific command, and to a specific CI context.
-
-    Note: We cannot hoist this higher in the command hierarchy because only one level of
-    subcommands are available at the time the context is created.
-    """
-
-    git_branch = ctx.obj["git_branch"]
-    git_revision = ctx.obj["git_revision"]
-    pipeline_start_timestamp = ctx.obj["pipeline_start_timestamp"]
-    ci_context = ctx.obj["ci_context"]
-    ci_job_key = ctx.obj["ci_job_key"] if ctx.obj.get("ci_job_key") else ci_context
-
-    sanitized_branch = git_branch.replace("/", "_")
-
-    # get the command name for the current context, if a group then prepend the parent command name
-    invoked_subcommand = ctx.invoked_subcommand
-    parent_command_path = ctx.command_path.replace(" ", "/") if ctx.command_path else None
-    cmd = f"{parent_command_path}/{invoked_subcommand}" if parent_command_path else invoked_subcommand
-
-    path_values = [
-        cmd,
-        ci_job_key,
-        sanitized_branch,
-        pipeline_start_timestamp,
-        git_revision,
-    ]
-
-    # check all values are defined
-    if None in path_values:
-        raise ValueError(f"Missing value required to render the report output prefix: {path_values}")
-
-    # join all values with a slash, and convert all values to string
-    return "/".join(map(str, path_values))
 
 
 # COMMANDS
@@ -154,16 +112,7 @@ def airbyte_ci(
     ctx.obj["ci_github_access_token"] = ci_github_access_token
     ctx.obj["ci_job_key"] = ci_job_key
     ctx.obj["pipeline_start_timestamp"] = pipeline_start_timestamp
-    ctx.obj["report_output_prefix"] = render_report_output_prefix(ctx)
-    if not show_dagger_logs:
-        dagger_log_dir = Path(f"{LOCAL_REPORTS_PATH_ROOT}/{ctx.obj['report_output_prefix']}")
-        dagger_log_dir.mkdir(parents=True, exist_ok=True)
-        dagger_log_path = Path(f"{dagger_log_dir}/dagger.log").resolve()
-        dagger_log_path.touch()
-        ctx.obj["dagger_logs_path"] = dagger_log_path
-        main_logger.info(f"Saving dagger logs to: {dagger_log_path}")
-    else:
-        ctx.obj["dagger_logs_path"] = None
+    ctx.obj["show_dagger_logs"] = show_dagger_logs
 
     if pull_request_number and ci_github_access_token:
         ctx.obj["pull_request"] = github.get_pull_request(pull_request_number, ci_github_access_token)
@@ -183,20 +132,6 @@ def airbyte_ci(
         main_logger.info(f"Pull Request Number: {pull_request_number}")
         main_logger.info(f"Pipeline Start Timestamp: {pipeline_start_timestamp}")
         main_logger.info(f"Modified Files: {ctx.obj['modified_files']}")
-
-    @ctx.call_on_close
-    def upload_dagger_logs() -> None:
-        if ctx.obj["dagger_logs_path"]:
-            if ctx.obj["is_local"]:
-                main_logger.info(f"Dagger logs saved to {ctx.obj['dagger_logs_path']}")
-            if ctx.obj["is_ci"]:
-                dagger_logs_gcs_key = f"{ctx.obj['report_output_prefix']}/dagger-logs.txt"
-                upload_to_gcs(
-                    ctx.obj["dagger_logs_path"], ctx.obj["ci_report_bucket_name"], dagger_logs_gcs_key, ctx.obj["ci_gcs_credentials"]
-                )
-                gcs_uri = f"gs://{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
-                public_url = f"https://storage.googleapis.com/{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
-                main_logger.info(f"Dagger logs saved to {gcs_uri}. Public URL: {public_url}")
 
 
 airbyte_ci.add_command(connectors)
