+++ tools/ci_connector_ops/ci_connector_ops/pipelines/utils.py
@@ -18,7 +18,7 @@ import anyio
 import asyncer
 import click
 import git
-from ci_connector_ops.pipelines import main_logger
+from ci_connector_ops.pipelines import consts, main_logger
 from ci_connector_ops.utils import get_all_released_connectors, get_changed_connectors
 from dagger import Config, Connection, Container, DaggerError, File, ImageLayerCompression, QueryError
 from google.cloud import storage
@@ -347,12 +347,71 @@ class DaggerPipelineCommand(click.Command):
             "If you're running this command for the first time the Dagger engine image will be pulled, it can take a short minute..."
         )
         try:
+            ctx.obj["report_output_prefix"] = self.render_report_output_prefix(ctx)
+            if not ctx.obj["show_dagger_logs"]:
+                dagger_log_dir = Path(f"{consts.LOCAL_REPORTS_PATH_ROOT}/{ctx.obj['report_output_prefix']}")
+                dagger_log_dir.mkdir(parents=True, exist_ok=True)
+                dagger_log_path = Path(f"{dagger_log_dir}/dagger.log").resolve()
+                dagger_log_path.touch()
+                ctx.obj["dagger_logs_path"] = dagger_log_path
+                main_logger.info(f"Saving dagger logs to: {dagger_log_path}")
+            else:
+                ctx.obj["dagger_logs_path"] = None
             pipeline_success = super().invoke(ctx)
             if not pipeline_success:
                 raise DaggerError(f"Dagger Command {command_name} failed.")
         except DaggerError as e:
             main_logger.error(f"Dagger Command {command_name} failed", exc_info=e)
             sys.exit(1)
+        finally:
+            if ctx.obj.get("dagger_logs_path"):
+                if ctx.obj["is_local"]:
+                    main_logger.info(f"Dagger logs saved to {ctx.obj['dagger_logs_path']}")
+                if ctx.obj["is_ci"]:
+                    dagger_logs_gcs_key = f"{ctx.obj['report_output_prefix']}/dagger-logs.txt"
+                    upload_to_gcs(
+                        ctx.obj["dagger_logs_path"], ctx.obj["ci_report_bucket_name"], dagger_logs_gcs_key, ctx.obj["ci_gcs_credentials"]
+                    )
+                    gcs_uri = f"gs://{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
+                    public_url = f"https://storage.googleapis.com/{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
+                    main_logger.info(f"Dagger logs saved to {gcs_uri}. Public URL: {public_url}")
+
+    @staticmethod
+    def render_report_output_prefix(ctx: click.Context) -> str:
+        """Render the report output prefix for any command in the Connector CLI.
+
+        The goal is to standardize the output of all logs and reports generated by the CLI
+        related to a specific command, and to a specific CI context.
+
+        Note: We cannot hoist this higher in the command hierarchy because only one level of
+        subcommands are available at the time the context is created.
+        """
+
+        git_branch = ctx.obj["git_branch"]
+        git_revision = ctx.obj["git_revision"]
+        pipeline_start_timestamp = ctx.obj["pipeline_start_timestamp"]
+        ci_context = ctx.obj["ci_context"]
+        ci_job_key = ctx.obj["ci_job_key"] if ctx.obj.get("ci_job_key") else ci_context
+
+        sanitized_branch = slugify(git_branch.replace("/", "_"))
+
+        # get the command name for the current context, if a group then prepend the parent command name
+        cmd = ctx.command_path.replace(" ", "/") if ctx.command_path else None
+
+        path_values = [
+            cmd,
+            ci_job_key,
+            sanitized_branch,
+            pipeline_start_timestamp,
+            git_revision,
+        ]
+
+        # check all values are defined
+        if None in path_values:
+            raise ValueError(f"Missing value required to render the report output prefix: {path_values}")
+
+        # join all values with a slash, and convert all values to string
+        return "/".join(map(str, path_values))
 
 
 async def execute_concurrently(steps: List[Callable], concurrency=5):
