+++ tools/ci_connector_ops/tests/test_pipelines/test_utils.py
@@ -4,7 +4,7 @@
 from unittest import mock
 
 import pytest
-from ci_connector_ops.pipelines.commands import airbyte_ci
+from ci_connector_ops.pipelines import utils
 
 
 @pytest.mark.parametrize(
@@ -12,8 +12,7 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
     [
         (
             mock.MagicMock(
-                command_path="my_command_path",
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch",
                     "git_revision": "my_git_revision",
@@ -22,12 +21,11 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": None,
                 },
             ),
-            "my_command_path/my_subcommand/my_ci_context/my_branch/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_context/my_branch/my_pipeline_start_timestamp/my_git_revision",
         ),
         (
             mock.MagicMock(
-                command_path="my_command_path",
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch",
                     "git_revision": "my_git_revision",
@@ -36,12 +34,11 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": "my_ci_job_key",
                 },
             ),
-            "my_command_path/my_subcommand/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
         ),
         (
             mock.MagicMock(
-                command_path="my_command_path another_command",
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch",
                     "git_revision": "my_git_revision",
@@ -50,12 +47,11 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": "my_ci_job_key",
                 },
             ),
-            "my_command_path/another_command/my_subcommand/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
         ),
         (
             mock.MagicMock(
-                command_path=None,
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch",
                     "git_revision": "my_git_revision",
@@ -64,12 +60,11 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": "my_ci_job_key",
                 },
             ),
-            "my_subcommand/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_job_key/my_branch/my_pipeline_start_timestamp/my_git_revision",
         ),
         (
             mock.MagicMock(
-                command_path=None,
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch/with/slashes",
                     "git_revision": "my_git_revision",
@@ -78,12 +73,11 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": "my_ci_job_key",
                 },
             ),
-            "my_subcommand/my_ci_job_key/my_branch_with_slashes/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_job_key/my_branch_with_slashes/my_pipeline_start_timestamp/my_git_revision",
         ),
         (
             mock.MagicMock(
-                command_path=None,
-                invoked_subcommand="my_subcommand",
+                command_path="my command path",
                 obj={
                     "git_branch": "my_branch/with/slashes#and!special@characters",
                     "git_revision": "my_git_revision",
@@ -92,9 +86,9 @@ from ci_connector_ops.pipelines.commands import airbyte_ci
                     "ci_job_key": "my_ci_job_key",
                 },
             ),
-            "my_subcommand/my_ci_job_key/my_branch_with_slashesandspecialcharacters/my_pipeline_start_timestamp/my_git_revision",
+            "my/command/path/my_ci_job_key/my_branch_with_slashesandspecialcharacters/my_pipeline_start_timestamp/my_git_revision",
         ),
     ],
 )
 def test_render_report_output_prefix(ctx, expected):
-    assert airbyte_ci.render_report_output_prefix(ctx) == expected
+    assert utils.DaggerPipelineCommand.render_report_output_prefix(ctx) == expected
